refactor(rerun): move serve implementation to rerun_serve module
Flip module ownership so rerun_serve.py is canonical and sim2real_rerun_serve
is a backward-compat shim; update engine/tests imports and fix test_clients
missing Path/subprocess imports. Rate-limit agent rerun restarts to avoid
systemd start-limit-hit during live verification.
